refac: Segregate `_get_wholesale_metering_point_time_series` into own module
